Saving state used for osse. Modified namelist to hourly rw files instead of 6hourly with time window of 6hours. Updated other corresponding scripts. Added a new tool for grabbing nature files from hpss.

+++ fortran/run/run.ksh
@@ -1,3 +1,8 @@
+#!/bin/ksh
+
+instructions='''Before running this script you should run ./tools/getNatureFiles.ksh'''
+
+export ndate=/home/Donald.E.Lippi/bin/ndate
 
 # Check for existence of sim directories which contain bufr, log, nml, and setup files.
 # If they don't already exist, create them. Otherwise continues.
@@ -9,30 +14,35 @@ for dir in $dirs; do
 done
 
 # Compile the simulation code.
-src=../src
-cd $src
-make clean
-make
-cd -
+print -n "Compile drwsim? (y/n): ";read var; print ""
+if [[ $var == 'y' ]]; then
+   src=../src
+   cd $src
+   make clean
+   make
+   cd -
+   exefile=drwsim.x
+   ln -sf $src/${exefile} .
+fi
 
-exefile=drwsim.x
 namelist=namelist
 setup=setup.ksh
 csvfile=radar_sim_dev.csv
-ln -sf $src/${exefile} .
 ln -sf ../fix/${csvfile} .
 ln -sf ../fix/l2rwbufr.table.csv .
 ln -sf ../fix/l2rwbufr.table .
-
 
 # Create the jobs for simulating the observations now.
 # EXAMPLE:
 # FHMIN_GFS=6, FHMAX_GFS=18, FHOUT_GFS=1, FHOUT_GDAS=6
 # This example computes 3 simulated drw bufr files: one for 06z, 12z, and 18z.
-FHMIN_GFS=6   #start hour (6, 12, 18, 24, 30, 36, etc). 
-FHMAX_GFS=6  #end  hour (6, 12, 18, 24, 30, etc). FHMIN = FHMAX if simulating obs for single cyc (1 bufr file).
+#23,24,25,26, 27, 28, 29, 30 (September 23-30)
+# 0,24,48,72, 96,120,144,168
+#23,47,71,95,119,143,167,171
+FHMIN_GFS=168 #start hour (6, 12, 18, 24, 30, 36, etc). 
+FHMAX_GFS=171 #end  hour (6, 12, 18, 24, 30, etc). FHMIN = FHMAX if simulating obs for single cyc (1 bufr file).
 FHOUT_GFS=1   #increment hour by (1; hourly 4DEnVar). Probably don't want to change this.
-FHOUT_GDAS=6  #increment for gdas analysis (6; analysis every 6 hours @ 00, 06, etc.). Don't change.
+FHOUT_GDAS=1  #increment for gdas analysis (6; analysis every 6 hours @ 00, 06, etc.). Don't change.
 fcsthr=$FHMIN_GFS
 typeset -Z3 fcsthr
 typeset -Z3 fcsthrm3
@@ -46,36 +56,22 @@ typeset -Z3 fcsthrp3
 #The next 4 lines are for the NATURE RUN. There is only one NATURE, the config used
 #forecast only (gfs), started at 00z on 20180911, and is named NATURE-SDATE-EDATE.
 CDUMP="gfs"
-CYC="00"
-PDY="20180911"
-EXP="NATURE-2018091100-2018091800" #This should be the NATURE RUN
+valtime=2018092300 #this is where the experiment starts.
+valtime=`${ndate} +$FHMIN_GFS $valtime` #add the first x-hours given by FHMIN_GFS to original start time.
+PDY=`echo $valtime | cut -c 1-8`
+CYC=`echo $valtime | cut -c 9-10`
+EXP="NATURE-2018092300-2018100700" #This should be the NATURE RUN
 
 while [[ $fcsthr -le $FHMAX_GFS ]]; do
-   (( fcsthrm3=fcsthr - 3 ))
-   (( fcsthrm2=fcsthr - 2 ))
-   (( fcsthrm1=fcsthr - 1 ))
+   valcyc=`echo $valtime | cut -c 9-10`
+   valpdy=`echo $valtime | cut -c 1-8`
    fcsthrm0=$fcsthr
-   (( fcsthrp1=fcsthr + 1 ))
-   (( fcsthrp2=fcsthr + 2 ))
-   (( fcsthrp3=fcsthr + 3 ))
-   #filename="gfs.t00z.atmf${fcsthr}.nemsio" #Get filename change
-   filename3="${CDUMP}.t${CYC}z.atmf${fcsthrm3}.nemsio" #Get filename change
-   filename4="${CDUMP}.t${CYC}z.atmf${fcsthrm2}.nemsio" #Get filename change
-   filename5="${CDUMP}.t${CYC}z.atmf${fcsthrm1}.nemsio" #Get filename change
    filename6="${CDUMP}.t${CYC}z.atmf${fcsthrm0}.nemsio" #Get filename change
-   filename7="${CDUMP}.t${CYC}z.atmf${fcsthrp1}.nemsio" #Get filename change
-   filename8="${CDUMP}.t${CYC}z.atmf${fcsthrp2}.nemsio" #Get filename change
-   filename9="${CDUMP}.t${CYC}z.atmf${fcsthrp3}.nemsio" #Get filename change
+
    cp $namelist ./simnml/namelist.atmf${fcsthr}
-   sed -i "s/@filename3@/${filename3}/g" ./simnml/namelist.atmf${fcsthr}
-   sed -i "s/@filename4@/${filename4}/g" ./simnml/namelist.atmf${fcsthr}
-   sed -i "s/@filename5@/${filename5}/g" ./simnml/namelist.atmf${fcsthr}
    sed -i "s/@filename6@/${filename6}/g" ./simnml/namelist.atmf${fcsthr}
-   sed -i "s/@filename7@/${filename7}/g" ./simnml/namelist.atmf${fcsthr}
-   sed -i "s/@filename8@/${filename8}/g" ./simnml/namelist.atmf${fcsthr}
-   sed -i "s/@filename9@/${filename9}/g" ./simnml/namelist.atmf${fcsthr}
    sed -i "s/@CDUMP@/${CDUMP}/g"       ./simnml/namelist.atmf${fcsthr}
-   sed -i "s/@PDY@/${PDY}/g"           ./simnml/namelist.atmf${fcsthr}
+   sed -i "s/@PDY@/${valpdy}/g"        ./simnml/namelist.atmf${fcsthr}
    sed -i "s/@CYC@/${CYC}/g"           ./simnml/namelist.atmf${fcsthr}
    sed -i "s/@EXP@/${EXP}/g"           ./simnml/namelist.atmf${fcsthr}
 
@@ -88,4 +84,5 @@ while [[ $fcsthr -le $FHMAX_GFS ]]; do
    qsub ./simsetup/setup_atmf${fcsthr}.ksh
    #ksh ./simsetup/setup_atmf${fcsthr}.ksh
    (( fcsthr=fcsthr+${FHOUT_GDAS} ))
+   valtime=`${ndate} +$FHOUT_GFS $valtime` #now update by FHOUT_GFS.
 done
